requirements.txt
# pip requirements file for MARCO

django-picklefield>=0.2.0
